src/requirements.txt
RPi.GPIO
Encoder
logzero
